begin of clean-up, rename simcore.h to simdefs.h
To avoid confusing github, first rename simcore.h to simdefs.h
and introduce the new simcore.h later.
