+++ cromemcosim/srcsim/sim.h
@@ -82,6 +82,6 @@ extern void sleep_ms(int);
 #define USR_CPR	"Copyright (C) 2014-2024 by Udo Munk & " \
 		"2021-2022 by David McNaughton"
 
-#include "simcore.h"
+#include "simdefs.h"
 
 #endif
