+++ intelmdssim/srcsim/sim.h
@@ -59,6 +59,6 @@ extern void sleep_ms(int);
 #define USR_CPR	"Copyright (C) 2008-2024 by Udo Munk & " \
 		"2024 by Thomas Eberhardt"
 
-#include "simcore.h"
+#include "simdefs.h"
 
 #endif /* !SIM_INC */
