+++ z80sim/srcsim/sim.h.fast
@@ -45,6 +45,6 @@ extern void sleep_ms(int);
 #define USR_CPR	"Copyright (C) 20xx by XYZ"
 */
 
-#include "simcore.h"
+#include "simdefs.h"
 
 #endif
